:recycle: Extract withFragments functionality into superclasses
