+++ app/api/views/fragments.py
@@ -79,3 +79,36 @@ class WikiFragmentViewSet(BaseFragmentViewSet):
     serializer_class = WikiFragmentSerializer
     filter_fields = ('wiki',)
     fragment_group_field = 'wiki'
+
+
+class ModelViewSetWithFragments(viewsets.ModelViewSet):
+
+    def get_serializer(self, *args, **kwargs):
+        kwargs['with_fragments'] = bool(
+            strtobool(self.request.query_params.get('withFragments', 'no'))
+        )
+        return super().get_serializer(*args, **kwargs)
+
+
+class ModelSerializerWithFragments(serializers.ModelSerializer):
+
+    def __init__(self, *args, with_fragments=False, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.with_fragments = with_fragments
+
+    def to_representation(self, obj):
+        rep = super().to_representation(obj)
+        if not self.with_fragments:
+            return rep
+
+        serializer = self.fragment_serializer_class(many=True, read_only=True)
+        fragments = obj.fragments.all()
+
+        next_fragments = []
+        for fragment in fragments:
+            next_fragments.append(fragment)
+            if fragment.question:
+                break
+
+        rep['next_fragments'] = serializer.to_representation(next_fragments)
+        return rep
